// lyndon/src/utils/enums.ts
export enum Size {
  XS = 'xs',
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
  XL = 'xl',
  XXL = 'xxl',
  FULL = 'full',
}

export enum Spacing {
  XXS = 'xxs',
  XS = 'xs',
  SM = 'sm',
  MD = 'md',
  LG = 'lg',
  XL = 'xl',
  XXL = 'xxl',
}

/* eslint-disable @typescript-eslint/no-duplicate-enum-values */
export enum IconName {
  ARROW_BACK = 'arrow-back',
  ARROW_RIGHT = 'arrow-right',
  BLUESKY = 'bluesky',
  BLUESKY_ALT = 'bluesky-alt',
  BLUESKY_SQUARE = 'bluesky-square',
  CHAPTER = 'chapter',
  CHECKMARK = 'checkmark',
  CHEVRON_DOWN = 'chevron-down',
  CHEVRON_LEFT = 'chevron-left',
  CHEVRON_RIGHT = 'chevron-right',
  CLOSE = 'close',
  DOWNLOAD = 'download',
  DOWNLOAD_ALT = 'download-alt',
  DROPDOWN = 'dropdown',
  EMAIL = 'email',
  EMAIL_ALT = 'email-alt',
  EXTERNAL = 'external',
  FACEBOOK = 'facebook',
  FACEBOOK_ALT = 'facebook-alt',
  FACEBOOK_SQUARE = 'facebook-square',
  FILTER = 'filter',
  GRID = '2x2grid',
  HAMBURGER_MENU = 'hamburger-menu',
  HAMBURGER_FOUR = 'hamburger-four',
  HOME = 'home',
  INFORMATION = 'information',
  INSTAGRAM = 'instagram',
  INSTAGRAM_ALT = 'instagram-alt',
  LINKEDIN = 'linkedin',
  LINKEDIN_ALT = 'linkedin-alt',
  MINUS = 'minus',
  MINUS_CIRCLE = 'minus-circle',
  PEOPLE = 'people',
  PLAY = 'play',
  PAUSE = 'pause',
  PIN = 'pin',
  PLUS = 'plus',
  PLUS_GRAY = 'plus-gray',
  PLUS_CIRCLE = 'plus-circle',
  REDDIT = 'reddit',
  RSS = 'rss',
  SHARE = 'share',
  SEARCH = 'search',
  THREADS = 'threads',
  THREADS_ALT = 'threads-alt',
  THREADS_SQUARE = 'threads-square',
  TRENDING = 'trending',
  TWITTER = 'twitter',
  TWITTER_ALT = 'twitter-alt',
  TWITTER_SQUARE = 'twitter-square',
  UNION = 'union',
  VIDEO = 'video',
  YOUTUBE = 'youtube',
  YOUTUBE_ALT = 'youtube-alt',
  X = 'twitter',
  X_ALT = 'twitter-alt',
  X_SQUARE = 'twitter-square',
  GITHUB = 'github',
  MICROPHONE = 'microphone',
}

export enum IconColor {
  'PRIMARY-C' = 'primary-c',
  BLACK = 'black',
  'PRIMARY-A' = 'primary-a',
  BLUE = 'blue',
  CURRENT = 'current',
  'PRIMARY-B' = 'primary-b',
  'PRIMARY-B-500' = 'primary-b-500',
  GRAY = 'gray',
  TRANSPARENT = 'transparent',
  'PRIMARY-D' = 'primary-d',
  WHITE = 'white',
}